LPD-82451 SF, move to a separate util class
