+++ modules/apps/portal/portal-util-test/src/testIntegration/java/com/liferay/portal/util/test/FriendlyURLUtilTest.java
@@ -1,0 +1,137 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.portal.util.test;
+
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.petra.string.StringBundler;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.service.GroupLocalService;
+import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
+import com.liferay.portal.kernel.test.util.GroupTestUtil;
+import com.liferay.portal.kernel.test.util.RandomTestUtil;
+import com.liferay.portal.kernel.test.util.TestPropsValues;
+import com.liferay.portal.kernel.test.util.UserTestUtil;
+import com.liferay.portal.test.rule.Inject;
+import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
+import com.liferay.portal.util.FriendlyURLUtil;
+
+import org.junit.Assert;
+import org.junit.ClassRule;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+/**
+ * @author Dante Wang
+ */
+@RunWith(Arquillian.class)
+public class FriendlyURLUtilTest {
+
+	@ClassRule
+	@Rule
+	public static final LiferayIntegrationTestRule liferayIntegrationTestRule =
+		new LiferayIntegrationTestRule();
+
+	@Test
+	public void testFetchFriendlyURLGroupWithGroupFriendlyURL()
+		throws Exception {
+
+		_group = GroupTestUtil.addGroup();
+
+		Group group = FriendlyURLUtil.fetchFriendlyURLGroup(
+			TestPropsValues.getCompanyId(), _group.getFriendlyURL());
+
+		Assert.assertEquals(_group.getGroupId(), group.getGroupId());
+	}
+
+	@Test
+	public void testFetchFriendlyURLGroupWithNullFriendlyURL()
+		throws Exception {
+
+		Assert.assertNull(
+			FriendlyURLUtil.fetchFriendlyURLGroup(
+				TestPropsValues.getCompanyId(), null));
+	}
+
+	@Test
+	public void testFetchFriendlyURLGroupWithoutMatchingGroupOrUser()
+		throws Exception {
+
+		Assert.assertNull(
+			FriendlyURLUtil.fetchFriendlyURLGroup(
+				TestPropsValues.getCompanyId(),
+				StringPool.SLASH + RandomTestUtil.randomString()));
+	}
+
+	@Test
+	public void testFetchFriendlyURLGroupWithUserScreenName() throws Exception {
+		User user = UserTestUtil.addUser();
+
+		Group userGroup = user.getGroup();
+
+		_groupLocalService.updateFriendlyURL(
+			userGroup.getGroupId(),
+			StringPool.SLASH + RandomTestUtil.randomString());
+
+		Group group = FriendlyURLUtil.fetchFriendlyURLGroup(
+			TestPropsValues.getCompanyId(),
+			StringPool.SLASH + user.getScreenName());
+
+		Assert.assertEquals(userGroup.getGroupId(), group.getGroupId());
+	}
+
+	@Test
+	public void testParseGroupFriendlyURLWithDocumentPathPrefix()
+		throws Exception {
+
+		_group = GroupTestUtil.addGroup();
+
+		Group otherGroup = GroupTestUtil.addGroup();
+
+		Assert.assertEquals(
+			otherGroup.getFriendlyURL(),
+			FriendlyURLUtil.parseGroupFriendlyURL(
+				StringBundler.concat(
+					_group.getFriendlyURL(), "/documents/d",
+					otherGroup.getFriendlyURL(), StringPool.SLASH,
+					RandomTestUtil.randomString())));
+	}
+
+	@Test
+	public void testParseGroupFriendlyURLWithLayoutSuffix() throws Exception {
+		_group = GroupTestUtil.addGroup();
+
+		Assert.assertEquals(
+			_group.getFriendlyURL(),
+			FriendlyURLUtil.parseGroupFriendlyURL(
+				_group.getFriendlyURL() + StringPool.SLASH +
+					RandomTestUtil.randomString()));
+	}
+
+	@Test
+	public void testParseGroupFriendlyURLWithoutSuffix() throws Exception {
+		_group = GroupTestUtil.addGroup();
+
+		Assert.assertEquals(
+			_group.getFriendlyURL(),
+			FriendlyURLUtil.parseGroupFriendlyURL(_group.getFriendlyURL()));
+	}
+
+	@Test
+	public void testParseGroupFriendlyURLWithRoot() throws Exception {
+		Assert.assertNull(
+			FriendlyURLUtil.parseGroupFriendlyURL(StringPool.SLASH));
+	}
+
+	@DeleteAfterTestRun
+	private Group _group;
+
+	@Inject
+	private GroupLocalService _groupLocalService;
+
+}
